Adding a commit#show api endpoint.
Also reordering the Api and Client apps to expose the json version of
a commit. If Client routes are checked first, /commit/:id will match
both the html and json endpoints. If the Api app comes first then it
it will match /commit/:id.json if necessary, and will pass on routes
that look like /commit/:id and let the Client app handle them.